"worked around" structure misalignment problem in test suite
I disabled a check below, because I can not find the cause of the
misalignment. The problem is that pToken structure has a different
member alignment inside the runtime library then inside of this
program. I checked compiler options, but could not find the cause.
Should anyone have any insight, I'd really appreciate if you drop me
a line.