bugfix: copy&paste error lead to dangling if
this caused a very minor issue with re-formatting a RFC3164 date when the
message was invalidly formatted and had a colon immediately after the date.
This was in the code for some years (even v1 had it) and I think it never
had any effect at all in practice. Though, it should be fixed - but definitely
nothing to worry about.