experimentally altered "last message repeated n times" to include msg
This was suggested by David Lang, to help identify the message that was
repeated. A problem is that I do not have the expanded template at
hand when the "last ... times" message is generated. Spending much time
on this functionality is also probably not a good thing, as the whole
functionality will be overhauled (and once this is done we will not at
all have the output template at hand). So the approach is to use a single
field - here msg - and inlcude it as a notation of what was repeated. This
is far from being perfect, but eventually good enough. I will now wait
for feedback before going any further.