omfile suspend handling for non-dynafiles, also bugfixes
primarily bugs introduced by recent changes. We now also handle
static file names correctly, that was not the case before. We
now correctly reset the descriptor in the dynafile cache if
somthing goes wrong.

Keep in mind that reliablity of output is depending on the
reliability of the file system driver (the cifs driver returns OK,
but still loses data if it is disconnected for too-long).