Add the ability to actually run statements.
It now runs SQL statements given as templates. In this case, the
template is given on the configuration file and the core passes the
SQL statement correctly formatted to doAction. I still need to decide
how to structure this for having prepared statements (prepare them at
parseSelector time) and then make doAction to only bind arguments and
execute. It commits after each statement, which is awfully slow but
good enough for the moment.

Next step after that is have a buffer of arguments, and make doAction
store new data as it arrives, then run the statement only when the
buffer is almost full. Or something like that.