added some tests for IPv4-Only configuration
We are "abusing" the parser tests to test the TCP/UDP reception
handling. Thus we do not run the other tests through to an IPv4
only config. It acutally is not that parser that matters but rather
the hope to find some init issues when running with only a single
listening socket (in IPv6, we usually both have an IPv4 AND an IPv6
listening socket, what - as practice has shown - may hide listener
setup errors).