first complete implementation of loadable parser system
I have now done the necessary cleanup. Looks like everything is in place.
Unfortunately, I do not yet have any actual parser that is not built-in,
but I think we can postpone working on that when the first one appears.
I don't expect troubles in that case, but you never know ;)