somewhat enhanced test suite
also prepared test case for TLS testing, but not activated it
inside the build itself.
A note on TLS testing: the current testsuite (in git!) already contains
TLS test cases. However, getting these test cases correct is not simple.
That's not a problem with the code itself, but rater a problem with
synchronization in the test environment. So I have deciced to keep the
TLS tests in, but not yet actually utilize them. This is most probably
left as an excercise for future (devel) releases.