goal reached: epoll() driver for plain tcp syslog receiver available
introduced the ablity for netstream drivers to utilize an epoll interface
This offers increased performance and removes the select() FDSET size
limit from imtcp. Note that we fall back to select() if there is no
epoll netstream drivers. So far, an epoll driver has only been
implemented for plain tcp syslog, the rest will follow once the code
proves well in practice AND there is demand.