experimental commit: facility to generate template via C function
This was a test done to try to generate templates with C code, via
a new (potentially to-be-implemented) class of template modules. We have
a rough POC inside this code, and it showed around 5% or better speedup. So
it semms worth continuing in this direction.

Note that this experimental commit works correct, but does any
template in the form of

$template tpl,=somewhat

will lead to fixed template expansion based on the default file format.