bugfix/imuxsock: did not compile on platforms without SCM_CREDENTIALS
Note: we do rate-limiting by the pid and we obtain the pid via SCM_CREDENTIALS
socket options. So this patch effectively disables the (new) ratelimiting
capability. In a later patch, I'll see that I can provide a global ratelimiting
capability, which could always be used (an alternative may be using the tag, will
check this out as well).