changed testbench parser patterns for more usual system config
a problem in the current testbench is that it depends on the locally
assigned system name. We have changed the default to the most common
default value. However, in the long term we need to fix this (and will
do so via a preloaded DNS cache, which we do not yet have!)