added new test to check RepeatedMsgReduction (via TCP)
This is kind of a backport of a similar UDP-based test from v6. However,
we do not have everything required for UDP in the v5 test tool base, so
we mimic the test with tcp -- which is not bad at all, because the
code path for TCP is somewhat different (and thus once we merge this
patch into v6, it is a useful addition).