Fall back to epoll_create() if epoll_create1() is not available
epoll_create1() was introduced in Linux kernel 2.6.27. If rsyslog was
compiled on a newer kernel but run on a kernel older than 2.6.27, remote
syslog fails. Apply a runtime check for epoll_create1() and fall back to
epoll_create() in this case.

Bug-Debian: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=617996