cleanup: remove omdbalerting
someone claimed to write a module with this functionality, if a
template were provided. And indeed, it was provided but never
anything happend (at least nothing was contributed back...). Removing
this silly template now. If someone really intends to take up on it,
get a copy from older git versions.