bugfix: imtcp could cause hang during reception
this also applied to other users of core file tcpsrv.c, but imtcp was
by far the most prominent and widely-used, the rest rather exotic
(like imdiag)

NOTE: this patch is probably not 100% correct and may cause problems
if multiple tcpsrv.c users existing within a configuration. The next
step is to evaluate this and probably slightly change the worker thread
startup/shutdown location.