bugfix: delayble source could block action queue, ...
even if there was a disk queue associated with it. The root cause of this
problem was that it makes no sense to delay messages once they arrive in the
action queue - the "input" that is being held in that case is the main
queue worker, what makes no sense.
Thanks to Marcin for alerting us on this problem and providing
instructions to reproduce it.