bugfix: omusrsmsg incorrect return state & config warning handling
During config file processing, Omusrmsg often incorrectly returned a
warning status, even when no warning was present (caused by
uninitialized variable). Also, the core handled warning messages
incorrectly, and treated them as errors. As a result, omusrmsg
(most often) could not properly be loaded. Note that this only
occurs with legacy config action syntax. This was a regression
caused by an incorrect merge in to the 6.3.x codebase.
Thanks to Stefano Mason for alerting us of this bug.