imuxsock: remove incorrect socket option call
SCM_CREDENTIALS is not a socket option; its value is usually
0x2 which on most archs corresponds to socket option
SO_REUSEADDR, but on mips-arch there is no socket option
with value 0x2 and SO_REUSEADDR = 0x4, so the call will fail
with ENOPROTOOPT 'Protocol not available'; skip it.
There does not seem any other special setsockopt call is
needed anyway, besides the above.

In addition Jonny Törnbom commented:
SCM_CREDENTIALS is a control message type, not a socket option, so using
setsockopt(...SCM_CREDENTIALS...) is potentially dangerous and wrong and
should be deleted from the code. (SCM_CREDENTIALS is used in conjuction
with SO_PASSCRED which is the socket option to use.)