Implemented different grammar for pri filters
The method introduced by the previous commit caused issues in v7 as it
was too generic. Most importantly, it permitted simple words (like "stop")
to (invalidly) be detected as pri filters. Now, the grammar is close
to the initial one, and each pri filter must at least have a comma or
a period inside it, which does not conflict with simple words.