bugfix: imuxsock truncated head of received message
This happened only under some circumstances. Thanks to Marius
Tomaschwesky, Florian Piekert and Milan Bartos for their help in
solving this issue.

Note that Milan sent a patch, which solved the problem under some
conditions (can be found somewhat earlier in git history). However,
it did not handle the original root cause, and so did not fix the
problem always. The main reason was that the date parser modifies
the provided lenghts. That part of the API was not properly
used by the caller (who decremented the length once more).